Add REUSE.toml files
REUSE.toml files are read by reuse to complement or override the
copyright and licensing information found in file.

The use of REUSE.toml files was introduced in REUSE version 3.1.0a1.
This reuse version is compatible with reuse specification
version 3.2 [1].
With this commit's files,
* The SPDX document generated by reuse spdx conforms to SPDX 2.3,
* The reuse lint command reports that the Qt project is reuse
compliant.

In order to be reuse compliant all the licenses present in the
LICENSES directory must be used. The license file not used
is removed.

[1]: https://reuse.software/spec-3.2/

Task-number: QTBUG-124453
Task-number: QTBUG-125211
Pick-to: 6.8
Change-Id: Ia489aa5b7d0d51024d8013aaa4b5d389492233e8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>