Fix and improve token request error reporting
From QAbstractOAuth::Error documentation:
NetworkError - Failed to connect to the server.
ServerError - The server answered the request with an error,
              or its response was not successfully received
The reply handler's usage of these error codes didn't match
these criteria. Notably it used QNetworkReply::error() function,
which mixes network, protocol, and HTTP status errors.

As per OAuth RFC, it is common for the authorization server to
respond with an HTTP error status like 400 Bad Request in case
something is wrong either on the client-side or on the server-side.

This commit makes a distinction between network/protocol errors,
and errors where the server responded with an HTTP status (or
errors which can be otherwise considered server-errors, like missing
content-type header). Network/protocol errors are NetworkErrors,
and other errors are ServerErrors.

As a related change, one qCWarning is removed as redundant, because
in practice tokenRequestErrorOccurred() handler also prints
a warning. Also as a related change, empty JSON object
branch was missing an error signal emission which is now added.

Note: Picking this to branches earlier than 6.6 isn't particularly
useful, because token request error reporting was introduced in 6.6.
Prior to 6.6, an error results only in a qCWarning print, not
an error signal emission.

Amends: 32f29d3e227da206f262efa055d1cac895855a98

[ChangeLog][QOAuthHttpServerReplyHandler] Make a better
distinction between NetworkErrors and ServerErrors with token
requests.

Pick-to: 6.8
Change-Id: I3ea8ca30da3e82867628ae5a90e2aa84e435d66d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>