Add .gitignore file
It's rather annoying to have all the CMakeLists.txt.user* files shows
in the output of `git status`, so maybe it's time to introduce a proper
.gitignore file to the module.

This commit takes the file from qtbase as a reference, and drops
everything that seems to be qtbase-specific.
It also updates the .gitattributes file to make sure that the
.gitignore file would not be a part of the packaged sources.
The REUSE.toml file is updated to make the CI happy, the change is
similar to how it is done in other submodules.

Pick-to: 6.9
Change-Id: I20ef7ef6d63042bfc06a60b65d47a1037814410f
Reviewed-by: Magdalena Stojek <magdalena.stojek@qt.io>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>