Rename isPolling property to polling
To comply with API naming guideline. In this context
the 'polling' can be considered an adjective/state, and
the getter is prefixed with 'is', while the name of the
property (and its change signal) is not prefixed.

https://wiki.qt.io/API_Design_Principles
See Naming Boolean Getters, Setters, and Properties

Resulted from API review.

Amends: 164e2d897fb7c3a60b518985774a4faa360ba2c9

Pick-to: 6.9
Change-Id: Iaf190edaf735ea32ca0ad66fe00a77a57bb037b4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>