QtNetworkAuth: fix QT_NO_HTTP checks
The QT_NO_HTTP definition is only available after including
QtNetwork/qtnetwork-config.h. However, the code was doing this check
before including any header.
As a result, the checks could be ignored if such header is the first
in the list of includes, or if nothing else includes qtnetwork-config.h
before it.

This patch makes sure that the QT_NO_HTTP definition is always available
before doing the check.

It is inspired by Qt 6.9 API review comments, but the problem exists in
earlier branches as well, so picking to all active Qt 6 branches.

Pick-to: 6.9 6.8 6.5
Change-Id: I7ddee5d41519c24a68df2e2cb956566cfe1d3619
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>